bugfix: fixed glitches with File and Folder selectors
- Folder selector needs a slash in the end of the path, otherwise it goes to the parent path. This happened in the Create Hardfile and Disk Swapper panels.
- File selector in Floppy and Quickstart panels would show incorrect directory contents, if it started with a filename/path that could not be resolved

+++ src/osdep/gui/PanelFloppy.cpp
@@ -73,7 +73,7 @@ public:
 	
 	int getNumberOfElements() override
 	{
-		return types.size();
+		return static_cast<int>(types.size());
 	}
 
 	std::string getElementAt(int i) override
@@ -113,7 +113,7 @@ public:
 
 	int getNumberOfElements() override
 	{
-		return types.size();
+		return static_cast<int>(types.size());
 	}
 
 	std::string getElementAt(int i) override
